Adicao da funcao save

+++ libft/ft_print_u.c
@@ -45,7 +45,7 @@ void		ft_print_u(t_ptf *parms)
 	p = 0;
 	temp = NULL;
 	parms->v_p = va_arg(parms->ap, void *);
-	parms->v_str = ft_itoa_base((unsigned int)parms->v_p, DEC_BASE);
+	parms->v_str = ft_itoa_base((uintmax_t)parms->v_p, DEC_BASE);
 	parms->len_c = ft_strlen(parms->v_str);
 	ft_aux(parms, temp);
 	p = parms->precision;
